requirements.txt
appdirs==1.4.0
Django==1.9
django-crispy-forms==1.6.1
django-formtools==2.0
django-simple-captcha==0.4.6
httplib2==0.9.2
MySQL-python==1.2.5
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
six==1.10.0
